add resources implement gameclear condition implement health bar

+++ 171018-BarrageShooting/WinAPI_Template/MapObject.cpp
@@ -17,6 +17,6 @@ void MapObject::Update()
 
 void MapObject::Render()
 {
-    m_imgBody->Render(g_hDC, m_dPos.x - GetSize().w * 0.5f, m_dPos.y - GetSize().h * 0.5f, GetSize().w, GetSize().h);
+    m_imgBody->Render(g_hDC, (int)(m_dPos.x - GetSize().w * 0.5f), (int)(m_dPos.y - GetSize().h * 0.5f), GetSize().w, GetSize().h);
     GameObject::Render();
 }
